Got it nearly functional
It distinguishes between new and active files and can make
a call to the API. Crudely, I have it appending the audio to the page without styling
just to make sure it works. Next up is storing the audio, displaying it to the user
in a nice way, and adding an export option!